# Improve influx parser performance