--- a/tremor-influx/Cargo.toml
+++ b/tremor-influx/Cargo.toml
@@ -11,3 +11,8 @@ simd-json = "0.3"
 
 [dev-dependencies]
 pretty_assertions = "0.6"
+criterion = "0.3"
+
+[[bench]]
+name = "decoder"
+harness = false