--- a/tremor-influx/Cargo.toml
+++ b/tremor-influx/Cargo.toml
@@ -8,6 +8,8 @@ edition = "2018"
 
 [dependencies]
 simd-json = "0.3"
+lexical = "5.2.0"
+
 
 [dev-dependencies]
 pretty_assertions = "0.6"