--- a/tremor-influx/Cargo.toml
+++ b/tremor-influx/Cargo.toml
@@ -12,6 +12,7 @@ simd-json = "0.3"
 [dev-dependencies]
 pretty_assertions = "0.6"
 criterion = "0.3"
+snmalloc-rs = "0.2"
 
 [[bench]]
 name = "decoder"