Performance fixes for influx decoder to fix #82 

Note every commit includes benchmark results and changes to the prior one.

Final results:
```
Gnuplot not found, using plotters backend
simple-value            time:   [376.98 ns 377.26 ns 377.52 ns]                         
                        change: [-55.627% -55.569% -55.518%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild

int-value               time:   [233.04 ns 233.19 ns 233.34 ns]                      
                        change: [-47.584% -47.507% -47.435%] (p = 0.00 < 0.05)
                        Performance has improved.

str-value               time:   [302.36 ns 302.43 ns 302.50 ns]                      
                        change: [-58.246% -58.202% -58.159%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) low mild
  4 (4.00%) high mild

escaped-value           time:   [364.08 ns 364.35 ns 364.67 ns]                          
                        change: [-52.159% -52.084% -52.004%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) high severe
```

They clearly hit the 30% mark set in #82.